+++ client/src/Components/Options.tsx
@@ -1,9 +1,9 @@
 "use client"
 
-import { useState } from "react"
-import { MoreVertical, User, Users, UserPlus } from "lucide-react"
-import { Button } from "@/Components/ui/button"
-import { Popover, PopoverContent, PopoverTrigger } from "@/Components/ui/popover"
+import { useState } from "react";
+import { MoreVertical, User, Users, UserPlus } from "lucide-react";
+import { Button } from "@/Components/ui/button";
+import { Popover, PopoverContent, PopoverTrigger } from "@/Components/ui/popover";
 import { useNavigate } from "react-router-dom";
 // import toast from "react-hot-toast"
 // import { useAuthStore } from "../store/useAuthStore";