+++ server/src/routes/message.route.js
@@ -1,6 +1,6 @@
 import express from "express";
 import { authenticate } from "../middleware/auth.middleware.js";
-import { getUsers, getMessage, sendMessage } from "../controllers/message.controller.js";
+import { getUsers, getMessage, sendMessage, deleteMessage } from "../controllers/message.controller.js";
 
 const router = express.Router();
 
@@ -9,5 +9,7 @@ router.get("/users", authenticate, getUsers);
 router.get("/:id", authenticate, getMessage);
 
 router.post("/send/:id", authenticate, sendMessage);
+
+router.delete("/delete/:id", authenticate, deleteMessage);
 
 export default router;