+++ History.md
@@ -1,7 +1,16 @@
+# 1.2.1 / 2025-03-07
+
+*Added version in api- (currently at v1)
+*Added delete user from admin panel feature
+*Added delete message for user
+*Added Copy message feature
+*Updated user is active or not feature in admin dashboard
+*Fixed minor bugs for improved stability
+
 # 1.2.0 / 2025-03-05
 
 *Added join room feature
-*Update admin dashboard has been modefied
+*Updated admin dashboard has been modefied
 *Fixed creating room functnality
 *Fixed Admin login bug 
 *Fixed minor bugs for improved stability